Users can not be deleted with foreign key violation #1942

+++ src/Messages/Entity/Message.php
@@ -249,6 +249,8 @@ class Message extends Entity
      */
     public function delete(): bool
     {
+        global $gSettingsManager;
+
         $this->db->startTransaction();
 
         $msgId = (int) $this->getValue('msg_id');
@@ -257,9 +259,11 @@ class Message extends Entity
             // first delete attachments files and the database entry
             $attachments   = $this->getAttachmentsInformation();
 
-            foreach ($attachments as $attachment) {
-                // delete attachment in file system
-                FileSystemUtils::deleteFileIfExists(ADMIDIO_PATH . FOLDER_DATA . '/messages_attachments/' . $attachment['admidio_file_name']);
+            if ($gSettingsManager->getBool('mail_save_attachments')) {
+                foreach ($attachments as $attachment) {
+                    // delete attachment in file system
+                    FileSystemUtils::deleteFileIfExists(ADMIDIO_PATH . FOLDER_DATA . '/messages_attachments/' . $attachment['admidio_file_name']);
+                }
             }
 
             $sql = 'DELETE FROM '.TBL_MESSAGES_ATTACHMENTS.'
